CHANGE: to_path now uses a reference rather than a DevTreeIndexNode
In the journey to understand Rust ownership and lifetimes I migrated
to_path to use references. Which led to use recursion rather than
loop to keep lifetimes under control.

More generally, I would love to find a way to avoid creation of
temporary objects (node.parent() creates a new one rather than
giving a reference to it. As most of the time values won't change
it seems a better approach but unsure at this stage of Rust learning